+++ RELEASES.md
@@ -16,6 +16,7 @@
 - `InputMap` now uses a collected `InputStreams` struct in all of its methods, and input methods are now optional
 - `InputManagerPlugin` now works even if some input stream resources are missing
 - added the `input_pressed` method to `InputMap`, to check if a single input is pressed
+- renamed `InputMap::assign_gamepad` to `InputMap::set_gamepad` for consistency and clarity (it does not uniquely assign a gamepad)
 
 ### Bug fixes